requirements.txt
web3
yaspin
colorama
pyyaml